destory needs check

+++ src/components/ebay-menu/component.js
@@ -131,6 +131,8 @@ module.exports = assign({}, menuUtils, {
             this.rovingTabindex.destroy();
             scrollKeyPreventer.remove(this.contentEl);
         }
-        this.destroyTypeahead();
+        if (this.destroyTypeahead) {
+            this.destroyTypeahead();
+        }
     }
 });
